Add create router tests
